break out filter bar
